generate methods
