+++ fastnode.go
@@ -5,9 +5,7 @@ package bart
 
 import (
 	"iter"
-	"net/netip"
 
-	"github.com/gaissmai/bart/internal/art"
 	"github.com/gaissmai/bart/internal/bitset"
 	"github.com/gaissmai/bart/internal/lpm"
 )
@@ -316,162 +314,5 @@ func (n *fastNode[V]) allot(idx uint8, oldValPtr, valPtr *V) {
 		// left:  idx*2
 		// right: (idx*2)+1
 		stack = append(stack, idx<<1, (idx<<1)+1)
-	}
-}
-
-// insertAtDepth inserts a network prefix and its associated value into the
-// trie starting at the specified byte depth.
-//
-// The function walks the prefix address from the given depth and inserts the value either directly into
-// the node's prefix table or as a compressed leaf or fringe node. If a conflicting leaf or fringe exists,
-// it is pushed down via a new intermediate node. Existing entries with the same prefix are overwritten.
-func (n *fastNode[V]) insertAtDepth(pfx netip.Prefix, val V, depth int) (exists bool) {
-	ip := pfx.Addr() // the pfx must be in canonical form
-	octets := ip.AsSlice()
-	lastOctetPlusOne, lastBits := lastOctetPlusOneAndLastBits(pfx)
-
-	// find the proper trie node to insert prefix
-	// start with prefix octet at depth
-	for ; depth < len(octets); depth++ {
-		octet := octets[depth]
-
-		// last masked octet: insert/override prefix/val into node
-		if depth == lastOctetPlusOne {
-			return n.insertPrefix(art.PfxToIdx(octet, lastBits), val)
-		}
-
-		kidAny, ok := n.getChild(octet)
-		// reached end of trie path ...
-		if !ok {
-			// insert prefix path compressed as leaf or fringe
-			if isFringe(depth, pfx) {
-				return n.insertChild(octet, newFringeNode(val))
-			}
-			return n.insertChild(octet, newLeafNode(pfx, val))
-		}
-
-		// kid is node or leaf at addr
-		switch kid := kidAny.(type) {
-		case *fastNode[V]:
-			n = kid // descend down to next trie level
-
-		case *leafNode[V]:
-			// reached a path compressed prefix
-			// override value in slot if prefixes are equal
-			if kid.prefix == pfx {
-				kid.value = val
-				// update
-				return true
-			}
-
-			// create new node
-			// push the leaf down
-			// insert new child at current leaf position (addr)
-			// descend down, replace n with new child
-			newNode := new(fastNode[V])
-			_ = newNode.insertAtDepth(kid.prefix, kid.value, depth+1)
-			_ = n.insertChild(octet, newNode)
-			n = newNode
-
-		case *fringeNode[V]:
-			// reached a path compressed fringe
-			// override value in slot if pfx is a fringe
-			if isFringe(depth, pfx) {
-				kid.value = val
-				// exists
-				return true
-			}
-
-			// create new node
-			// push the fringe down, it becomes a default route (idx=1)
-			// insert new child at current leaf position (addr)
-			// descend down, replace n with new child
-			newNode := new(fastNode[V])
-			_ = newNode.insertPrefix(1, kid.value)
-			_ = n.insertChild(octet, newNode)
-			n = newNode
-
-		default:
-			panic("logic error, wrong node type")
-		}
-	}
-
-	panic("unreachable")
-}
-
-// purgeAndCompress performs bottom-up compression of the trie by removing
-// empty nodes and converting sparse branches into compressed leaf/fringe nodes.
-// It unwinds the provided stack of parent nodes, checking each level for
-// compression opportunities based on child count and prefix distribution.
-//
-// The compression may convert:
-//   - Nodes with single prefix into leafNode (path compression)
-//   - Nodes at maxDepth-1 with single prefix into fringeNode
-//   - Empty intermediate nodes are removed entirely
-//
-// The reconstruction of prefixes for compressed nodes uses the traversal
-// path stored in octets and the parent's depth information.
-func (n *fastNode[V]) purgeAndCompress(stack []*fastNode[V], octets []uint8, is4 bool) {
-	// unwind the stack
-	for depth := len(stack) - 1; depth >= 0; depth-- {
-		parent := stack[depth]
-		octet := octets[depth]
-
-		pfxCount := n.prefixCount()
-		childCount := n.childCount()
-
-		switch {
-		case pfxCount == 0 && childCount == 0:
-			// just delete this empty node from parent
-			parent.deleteChild(octet)
-
-		case pfxCount == 0 && childCount == 1:
-			addr, _ := n.children.FirstSet() // single child must be first child
-			kidAny := *n.children.items[addr]
-
-			switch kid := kidAny.(type) {
-			case *fastNode[V]:
-				// fast exit, we are at an intermediate path node
-				// no further delete/compress upwards the stack is possible
-				return
-			case *leafNode[V]:
-				// just one leaf, delete this node and reinsert the leaf above
-				parent.deleteChild(octet)
-
-				// ... (re)insert the leaf at parents depth
-				parent.insertAtDepth(kid.prefix, kid.value, depth)
-			case *fringeNode[V]:
-				// just one fringe, delete this node and reinsert the fringe as leaf above
-				parent.deleteChild(octet)
-
-				// rebuild the prefix with octets, depth, ip version and addr
-				// depth is the parent's depth, so add +1 here for the kid
-				fringePfx := cidrForFringe(octets, depth+1, is4, addr)
-
-				// ... (re)reinsert prefix/value at parents depth
-				parent.insertAtDepth(fringePfx, kid.value, depth)
-			}
-
-		case pfxCount == 1 && childCount == 0:
-			// just one prefix, delete this node and reinsert the idx as leaf above
-			parent.deleteChild(octet)
-
-			// get prefix/val back from idx ...
-			idx, _ := n.prefixes.FirstSet() // single idx must be first bit set
-			val := *n.prefixes.items[idx]
-
-			// ... and octet path
-			path := stridePath{}
-			copy(path[:], octets)
-
-			// depth is the parent's depth, so add +1 here for the kid
-			pfx := cidrFromPath(path, depth+1, is4, idx)
-
-			// ... (re)insert prefix/value at parents depth
-			parent.insertAtDepth(pfx, val, depth)
-		}
-
-		// climb up the stack
-		n = parent
 	}
 }
