+++ litenode.go
@@ -5,9 +5,7 @@ package bart
 
 import (
 	"iter"
-	"net/netip"
 
-	"github.com/gaissmai/bart/internal/art"
 	"github.com/gaissmai/bart/internal/bitset"
 	"github.com/gaissmai/bart/internal/lpm"
 	"github.com/gaissmai/bart/internal/sparse"
@@ -247,171 +245,4 @@ func (n *liteNode[V]) lookupIdx(idx uint8) (top uint8, _ V, ok bool) {
 func (n *liteNode[V]) lookup(idx uint8) (_ V, ok bool) {
 	_, _, ok = n.lookupIdx(idx)
 	return
-}
-
-// insertAtDepth inserts a network prefix and its associated value into the
-// trie starting at the specified byte depth.
-//
-// The function traverses the prefix address from the given depth and inserts
-// the prefix either directly into the node's prefix table or as a compressed
-// leaf or fringe node. If a conflicting leaf or fringe exists, it creates
-// a new intermediate node to accommodate both entries.
-//
-// Parameters:
-//   - pfx: The network prefix to insert (must be in canonical form)
-//   - depth: The current depth in the trie (0-based byte index)
-//
-// Returns true if a prefix already existed and was updated, false for new insertions.
-func (n *liteNode[V]) insertAtDepth(pfx netip.Prefix, zero V, depth int) (exists bool) {
-	ip := pfx.Addr() // the pfx must be in canonical form
-	octets := ip.AsSlice()
-	lastOctetPlusOne, lastBits := lastOctetPlusOneAndLastBits(pfx)
-
-	// find the proper trie node to insert prefix
-	// start with prefix octet at depth
-	for ; depth < len(octets); depth++ {
-		octet := octets[depth]
-		// last masked octet: insert/override prefix/val into node
-		if depth == lastOctetPlusOne {
-			return n.insertPrefix(art.PfxToIdx(octet, lastBits), zero)
-		}
-
-		// reached end of trie path ...
-		if !n.children.Test(octet) {
-			// insert prefix path compressed as leaf or fringe
-			if isFringe(depth, pfx) {
-				return n.insertChild(octet, newFringeNode(zero))
-			}
-			return n.insertChild(octet, newLeafNode(pfx, zero))
-		}
-
-		// ... or descend down the trie
-		kid := n.mustGetChild(octet)
-
-		// kid is node or leaf or fringe at addr
-		switch kid := kid.(type) {
-		case *liteNode[V]:
-			n = kid // descend down to next trie level
-
-		case *leafNode[V]:
-			// reached a path compressed prefix
-			if kid.prefix == pfx {
-				// exists
-				return true
-			}
-
-			// create new node
-			// push the leaf down
-			// insert new child at current leaf position (addr)
-			// lite has no values
-			// descend down, replace n with new child
-			newNode := new(liteNode[V])
-			newNode.insertAtDepth(kid.prefix, zero, depth+1)
-
-			n.insertChild(octet, newNode)
-			n = newNode
-
-		case *fringeNode[V]:
-			// reached a path compressed fringe
-			if isFringe(depth, pfx) {
-				// exists
-				return true
-			}
-
-			// create new node
-			// push the fringe down, it becomes a default route (idx=1)
-			// insert new child at current leaf position (addr)
-			// lite has no values
-			// descend down, replace n with new child
-			newNode := new(liteNode[V])
-			newNode.insertPrefix(1, zero)
-
-			n.insertChild(octet, newNode)
-			n = newNode
-
-		default:
-			panic("logic error, wrong node type")
-		}
-	}
-
-	panic("unreachable")
-}
-
-// purgeAndCompress performs bottom-up compression of the trie by removing
-// empty nodes and converting sparse branches into compressed leaf/fringe nodes.
-//
-// The function unwinds the provided stack of parent nodes, checking each level
-// for compression opportunities based on child count and prefix distribution.
-// It may convert:
-//   - Nodes with a single prefix into leafNode (path compression)
-//   - Nodes at lastOctet with a single prefix into fringeNode
-//   - Empty intermediate nodes are removed entirely
-//
-// Parameters:
-//   - stack: Array of parent nodes to process during unwinding
-//   - octets: The path of octets taken to reach the current position
-//   - is4: True for IPv4 processing, false for IPv6
-func (n *liteNode[V]) purgeAndCompress(stack []*liteNode[V], octets []uint8, is4 bool) {
-	var zero V
-
-	// unwind the stack
-	for depth := len(stack) - 1; depth >= 0; depth-- {
-		parent := stack[depth]
-		octet := octets[depth]
-
-		childCount := n.childCount()
-
-		switch {
-		case n.prefixes.count == 0 && childCount == 0:
-			// just delete this empty node from parent
-			parent.deleteChild(octet)
-
-		case n.prefixes.count == 0 && childCount == 1:
-			switch kid := n.children.Items[0].(type) {
-			case *liteNode[V]:
-				// fast exit, we are at an intermediate path node
-				// no further delete/compress upwards the stack is possible
-				return
-			case *leafNode[V]:
-				// just one leaf, delete this node and reinsert the leaf above
-				parent.deleteChild(octet)
-
-				// ... (re)insert the leaf at parents depth
-				parent.insertAtDepth(kid.prefix, zero, depth)
-			case *fringeNode[V]:
-				// just one fringe, delete this node and reinsert the fringe as leaf above
-				parent.deleteChild(octet)
-
-				// get the last octet back, the only item is also the first item
-				lastOctet, _ := n.children.FirstSet()
-
-				// rebuild the prefix with octets, depth, ip version and addr
-				// depth is the parent's depth, so add +1 here for the kid
-				fringePfx := cidrForFringe(octets, depth+1, is4, lastOctet)
-
-				// ... (re)reinsert prefix/value at parents depth
-				parent.insertAtDepth(fringePfx, zero, depth)
-			}
-
-		case n.prefixes.count == 1 && childCount == 0:
-			// just one prefix, delete this node and reinsert the idx as leaf above
-			parent.deleteChild(octet)
-
-			// get prefix back from idx ...
-			idx, _ := n.prefixes.FirstSet() // single idx must be first bit set
-
-			// ... and octet path
-			path := stridePath{}
-			copy(path[:], octets)
-
-			// depth is the parent's depth, so add +1 here for the kid
-			pfx := cidrFromPath(path, depth+1, is4, idx)
-
-			// ... (re)insert prefix/value at parents depth
-			parent.insertAtDepth(pfx, zero, depth)
-		}
-
-		// climb up the stack
-		n = parent
-	}
 }
